Scoreboard api (#67)
* Scoreboard API implemented

* TimeEpoch Fix

* capitalization

* EnumFixes

* fixed time

+++ EnoCore/Models/Database/Service.cs
@@ -9,9 +9,9 @@ namespace EnoCore.Models.Database
 #pragma warning disable CS8618
         public long Id { get; set; }
         public string Name { get; set; }
-        public int FlagsPerRound { get; set; }
-        public int NoisesPerRound { get; set; }
-        public int HavocsPerRound { get; set; }
+        public long FlagsPerRound { get; set; }
+        public long NoisesPerRound { get; set; }
+        public long HavocsPerRound { get; set; }
         public long ServiceStatsId { get; set; }
         public List<ServiceStats> ServiceStats { get; set; }
         public bool Active { get; set; }
